refactor: auth api 디렉토리 변경
