+++ src/apis/auth/auth.ts
@@ -1,6 +1,6 @@
-import { AccessToken, SignInData } from '../features/auth/types/data';
-import instance from '.';
+import instance from '..';
 import { AUTH_API } from '@/features/auth/shared/constants/api';
+import { AccessToken, SignInData } from './types/auth';
 
 const authApi = {
   /**
