+++ src/features/auth/hooks/useGoogleLogin.ts
@@ -2,7 +2,7 @@
 
 import { initializeApp } from 'firebase/app';
 import { GoogleAuthProvider, getAuth, signInWithPopup } from 'firebase/auth';
-import authApi from '@/apis/auth';
+import authApi from '@/apis/auth/auth';
 import { useAuthActions, useIsSignedIn } from '../store';
 
 const firebaseConfig = {
